Add singleton trait and update integrations to use it
