+++ inc/integrations/class-archiveless.php
@@ -7,29 +7,13 @@
 
 namespace WP_Irving\Integrations;
 
+use WP_Irving\Singleton;
+
 /**
  * Class to replicate Archiveless's query modifications.
  */
 class Archiveless {
-
-	/**
-	 * Class instance.
-	 *
-	 * @var null|self
-	 */
-	protected static $instance;
-
-	/**
-	 * Get class instance.
-	 *
-	 * @return self
-	 */
-	public static function instance() {
-		if ( ! isset( static::$instance ) ) {
-			static::$instance = new static();
-		}
-		return static::$instance;
-	}
+	use Singleton;
 
 	/**
 	 * Constructor for class.
