+++ inc/integrations/class-google-analytics.php
@@ -7,10 +7,13 @@
 
 namespace WP_Irving\Integrations;
 
+use WP_Irving\Singleton;
+
 /**
  * Class to integrate Google Analytics with Irving.
  */
 class Google_Analytics {
+	use Singleton;
 
 	/**
 	 * The option key for the integration.
@@ -27,36 +30,18 @@ class Google_Analytics {
 	private $options;
 
 	/**
-	 * Class instance.
-	 *
-	 * @var null|self
-	 */
-	protected static $instance;
-
-	/**
-	 * Get class instance.
-	 *
-	 * @return self
-	 */
-	public static function instance() {
-		if ( ! isset( static::$instance ) ) {
-			static::$instance = new static();
-		}
-		return static::$instance;
-	}
-
-	/**
 	 * Setup the singleton. Validate JWT is installed, and setup hooks.
 	 */
 	public function setup() {
 		// Retrieve any existing integrations options.
 		$this->options = get_option( 'irving_integrations' );
+
 		// Register settings fields for integrations.
 		add_action( 'admin_init', [ $this, 'register_settings_fields' ] );
+
 		// Filter the updated option values prior to submission.
 		add_filter( 'pre_update_option_irving_integrations', [ $this, 'group_and_format_options_for_storage' ] );
 	}
-
 
 	/**
 	 * Register settings fields for display.
